Revert "Use the flake registry for the dotfiles flake"
This was a cute idea, *but* it turns out that the hash this produces
in the nix store isn't stable between computers.

So n-1 repositories would be cursed with forever-dirtiness anyway, and
this makes writing GitHub actions harder. Wait for `nix flake` to
become a bit more powerful for a better solution (or contribute
something like `nix flake --override-input`).